Use gsub to solve general problem
This commit uses gsub to replae the opening and closing parentheses. It
is not as elegant an implementation as we thought, since the string "-1"
gets split by chars.

As a result the reduce block requires additional work before it can
properly calculate the total.

I think it would be better to leave it as chars and then use flow
control to assign the correct value to n within the block, but I
implemented the solution anyway for your review.